requirements.txt
librosa==0.9.2
matplotlib==3.6.3
mediapipe-silicon==0.8.11
numpy==1.23.5
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
scikit-learn==1.2.0
scipy==1.10.0
urllib3==1.26.14